+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## Dev
+* chore: use Object.hasOwn instead of hasOwnProperty
+
 ## 7.0.0
 * improvement: remove dependency on Cucumber runner, see all the details in [migration guide](), ([#136](https://github.com/vitalets/playwright-bdd/issues/136))
 * improvement: support custom parameter types in decorator steps, ([#112](https://github.com/vitalets/playwright-bdd/issues/112))